+++ microapps/ScalebarApp/app/src/main/java/com/arcgismaps/toolkit/scalebarapp/screens/MainScreen.kt
@@ -41,7 +41,6 @@ import com.arcgismaps.mapping.Viewpoint
 import com.arcgismaps.toolkit.geoviewcompose.MapView
 import com.arcgismaps.toolkit.scalebar.Scalebar
 import com.arcgismaps.toolkit.scalebar.ScalebarStyle
-import com.arcgismaps.toolkit.scalebar.ScalebarUnits
 
 @Composable
 fun MainScreen(modifier: Modifier) {
